requirements.txt
Django>=4.2
djangorestframework
pandas
openpyxl
numpy
python-dateutil
django-cors-headers